project/requirements.txt
mypy==0.770
django-stubs==1.5.0
